`Rmd` changes
Moved around the yaml header output.

Added a settings file for the whole repo, so made some changes to the code to reflect a file that actually includes the version number.

Cleaned whitespace issues, made some fixes to plotting functions to use the `coord_cartesian()` function to set X and Y limits.

Trying to move the pointer to the new Baconizing Outputs repo for pulling in all the ages output by Bacon.